Cart & css updates
Created the way i waant the products to be in the cart. Still need to ask abt how to do the thingy with the java variabel to the php math n stuff. Other than that, gave the amount of items a local storage so it saves (for when u go to next page n its saved how much u got in there). Also fixed pageProduct bug for now.