Working
Working:
Homepage, Productpage
Somewhat working:
Cart, payment/checkout page
Needs lots of work:
PurchaseInfo page.
I have the basics for most of the things done, though the only thing that needs to be added to the cart is the ability to add more than 1 type of product. In the checkout, its more of making it neater and stuff, but further than that, it works. PurchaseInfo needs alot of work but for now, it works for what it is. Really simple to just get it working. Future plan: Get the id via $_GET or via the $_SESSION stuff.