1/2 Fix unsetting
Fixed the paying unset. Now it calls the unset session when the submit button is clicked when paying. Still don't know how i'm gonna do the clear one. I thought abt doing a shortcut through a php doc where it just started and unset the session, and then went back to the cart.php but that didn't work.