More fixes
sorta fixed the clear issue of the cart ig. It now redirects to a page, and i wanted it to immediately send it to the cart again but it don't do that so i added a href so it redirects back. Also fixed issue with paying method where it already cleared the Session when on the payment screen without having filled in the form.