Cart n bugs
Alr so idk i did the session thing n it now works but only for 1 miniscule product. Also discovered a bug w the cart where if you click on the product page, and then the cart, it's already saved the info and not the amount cuz u didnt fill in the needed info (duh).